refactoring folders handlers
