+++ server/handlers/orgs/browse-add-tags-view.js
@@ -1,21 +1,20 @@
 'use strict';
 
-var helpers = require('./helpers.js');
+var helpers = require('../helpers.js');
 
 /**
 * Switch out redis --> postgres:
 * Build 'options' object, simplify
 */
 
-module.exports = (request, reply) => {
+module.exports = function (request, reply) {
   var permissions = helpers.getPermissions(request.auth.credentials);
   // request.getTags(function (err, allTags) {
   // });
   request.redis.HGET('tags', 'tags', (error, allTags) => {
     var parent_tags = JSON.parse(allTags);
-    // Define which of the tabs are selected --> 'orgs' or 'challenges' view
-    var view = { [request.route.path.split('/')[1]]: true };
-    var options = Object.assign({}, {view}, {parent_tags}, permissions);
+    var searchAll = 'orgs';
+    var options = Object.assign({}, {searchAll}, {parent_tags}, permissions);
     reply.view('browse/tags', options);
   });
 };
